Add consistency error and refactor utils
