+++ graphql-server/src/graphql/transaction/dataProviders.js
@@ -1,6 +1,8 @@
 // @flow
 import assert from 'assert'
-import {parseAdaValue, annotateNotFoundError, validate} from '../utils'
+import {parseAdaValue} from '../utils'
+import {annotateNotFoundError} from '../../utils/errors'
+import {validate} from '../../utils/validation'
 import E from '../../api/elasticHelpers'
 
 export const facadeTransaction = (source: any) => {
