+++ graphql-server/src/utils/common.js
@@ -1,0 +1,4 @@
+// @flow
+
+// TODO: type better or use other compose function
+export const compose = (...fns: any) => (x: any) => fns.reduceRight((y, f) => f(y), x)
